at least no errors

+++ SEP3_CSharp/DataTransferObjects/ItemDTOs.cs
@@ -5,7 +5,7 @@ public class ItemDTOs
     public int ItemId{get;set;}
     public string Name{get;set;}
     public string Description{get;set;}
-    public double Price{get;set;}
+    public float Price{get;set;}
     public int Quantity{get;set;}
     public int CategoryId{get;set;}
     public string CategoryName{get;set;}
@@ -14,7 +14,7 @@ public class ItemDTOs
     public ItemDTOs()
     {
     }
-    public ItemDTOs(int itemId, string name, string description, double price, int quantity, int categoryId,
+    public ItemDTOs(int itemId, string name, string description, float price, int quantity, int categoryId,
         string categoryName, string imageURL)
     {
         this.ItemId = itemId;
